fix variable types for entities so it matches in tables. new inital m…

+++ src/CiPeWorldCup.Application/Queries/GetMatchesForSpecificRound.cs
@@ -5,19 +5,19 @@ namespace CiPeWorldCup.Tests;
 
 public class GetMatchesForSpecificRound
 {
-    public Result<long> GetAllMatchesForSpecificRound(long numberOfParticipants, int roundNumber)
+    public Result<int> GetAllMatchesForSpecificRound(int numberOfParticipants, int roundNumber)
     {
         var validationResult = ValidateNumberOfParticipants.ValidateParticipants(numberOfParticipants);
         if (!validationResult.IsSuccess)
         {
-            return Result<long>.Fail(validationResult.Error);
+            return Result<int>.Fail(validationResult.Error);
         }
         var roundValidation = ValidateRoundNumber.ValidateRound(roundNumber, (int)numberOfParticipants);
         if (!roundValidation.IsSuccess)
         {
-            return Result<long>.Fail(roundValidation.Error);
+            return Result<int>.Fail(roundValidation.Error);
         }
-        long matchesPerRound = numberOfParticipants / 2;
-        return Result<long>.Ok(matchesPerRound);
+        int matchesPerRound = numberOfParticipants / 2;
+        return Result<int>.Ok(matchesPerRound);
     }
 }
