+++ src/CiPeWorldCup.Core/Validation/ValidateRoundNumber.cs
@@ -8,17 +8,17 @@ using System.Threading.Tasks;
 namespace CiPeWorldCup.Core.Validation;
 public class ValidateRoundNumber
 {
-    public static Result<long> ValidateRound(long roundNumber, long numberOfParticipants)
+    public static Result<int> ValidateRound(int roundNumber, int numberOfParticipants)
     {
         if (roundNumber < 1)
         {
-            return Result<long>.Fail("Round number must be at least 1");
+            return Result<int>.Fail("Round number must be at least 1");
         }
         if (roundNumber > numberOfParticipants - 1)
         {
-            return Result<long>.Fail($"Round number cannot exceed {numberOfParticipants - 1}");
+            return Result<int>.Fail($"Round number cannot exceed {numberOfParticipants - 1}");
         }
-        return Result<long>.Ok(roundNumber);
+        return Result<int>.Ok(roundNumber);
     }
 
 }
